backend/requirements.txt

flask
flask_cors
requests
fastapi
uvicorn
pydantic
python-dotenv
boto3
awswrangler
openai
pandas
